+++ config/tsconfig.json
@@ -1,0 +1,12 @@
+{
+  "compilerOptions": {
+    "module": "commonjs",
+    "target": "es2016",
+    "strict": true,
+    "moduleResolution": "node",
+    "outDir": "../built",
+  },
+  "include": [
+    "../src/**/*"
+  ]
+}
