+++ package.json
@@ -2,9 +2,12 @@
   "name": "api.codeyql.com",
   "version": "1.0.0",
   "description": "API for CodeYQL site",
-  "main": "app.js",
+  "main": "src/app.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "build": "tsc --project ./config/",
+    "build:watch": "npm run build -- --watch",
+    "start": "npm run build && node built/app.js"
   },
   "author": "Chris Rabl <chris@autovance.com>",
   "license": "MIT",
@@ -16,5 +19,12 @@
     "jsonwebtoken": "^8.3.0",
     "moment": "^2.22.2",
     "uuid": "^3.3.2"
+  },
+  "devDependencies": {
+    "@types/body-parser": "^1.17.0",
+    "@types/cors": "^2.8.4",
+    "@types/express": "^4.16.0",
+    "tslint": "^5.11.0",
+    "typescript": "^3.1.3"
   }
 }
